write dagger logs for all airbyte-ci command
