+++ tools/ci_connector_ops/ci_connector_ops/pipelines/commands/airbyte_ci.py
@@ -18,6 +18,7 @@ from ci_connector_ops.pipelines.utils import (
     get_modified_files_in_branch,
     get_modified_files_in_commit,
     get_modified_files_in_pull_request,
+    upload_to_gcs,
 )
 from github import PullRequest
 
@@ -51,6 +52,45 @@ def get_modified_files(
     return get_modified_files_in_branch(git_branch, git_revision, diffed_branch, is_local)
 
 
+def render_report_output_prefix(ctx: click.Context) -> str:
+    """Render the report output prefix for any command in the Connector CLI.
+
+    The goal is to standardize the output of all logs and reports generated by the CLI
+    related to a specific command, and to a specific CI context.
+
+    Note: We cannot hoist this higher in the command hierarchy because only one level of
+    subcommands are available at the time the context is created.
+    """
+
+    git_branch = ctx.obj["git_branch"]
+    git_revision = ctx.obj["git_revision"]
+    pipeline_start_timestamp = ctx.obj["pipeline_start_timestamp"]
+    ci_context = ctx.obj["ci_context"]
+    ci_job_key = ctx.obj["ci_job_key"] if ctx.obj.get("ci_job_key") else ci_context
+
+    sanitized_branch = git_branch.replace("/", "_")
+
+    # get the command name for the current context, if a group then prepend the parent command name
+    invoked_subcommand = ctx.invoked_subcommand
+    parent_command_path = ctx.command_path.replace(" ", "/") if ctx.command_path else None
+    cmd = f"{parent_command_path}/{invoked_subcommand}" if parent_command_path else invoked_subcommand
+
+    path_values = [
+        cmd,
+        ci_job_key,
+        sanitized_branch,
+        pipeline_start_timestamp,
+        git_revision,
+    ]
+
+    # check all values are defined
+    if None in path_values:
+        raise ValueError(f"Missing value required to render the report output prefix: {path_values}")
+
+    # join all values with a slash, and convert all values to string
+    return "/".join(map(str, path_values))
+
+
 # COMMANDS
 
 
@@ -71,6 +111,13 @@ def get_modified_files(
 @click.option("--ci-git-user", default="octavia-squidington-iii", envvar="CI_GIT_USER", type=str)
 @click.option("--ci-github-access-token", envvar="CI_GITHUB_ACCESS_TOKEN", type=str)
 @click.option("--ci-report-bucket-name", envvar="CI_REPORT_BUCKET_NAME", type=str)
+@click.option(
+    "--ci-gcs-credentials",
+    help="The service account to use during CI.",
+    type=click.STRING,
+    required=False,  # Not required for pre-release or local pipelines
+    envvar="GCP_GSM_CREDENTIALS",
+)
 @click.option("--ci-job-key", envvar="CI_JOB_KEY", type=str)
 @click.option("--show-dagger-logs/--hide-dagger-logs", default=False, type=bool)
 @click.pass_context
@@ -87,6 +134,7 @@ def airbyte_ci(
     ci_git_user: str,
     ci_github_access_token: str,
     ci_report_bucket_name: str,
+    ci_gcs_credentials: str,
     ci_job_key: str,
     show_dagger_logs: bool,
 ):  # noqa D103
@@ -101,11 +149,21 @@ def airbyte_ci(
     )
     ctx.obj["ci_context"] = ci_context
     ctx.obj["ci_report_bucket_name"] = ci_report_bucket_name
+    ctx.obj["ci_gcs_credentials"] = ci_gcs_credentials
     ctx.obj["ci_git_user"] = ci_git_user
     ctx.obj["ci_github_access_token"] = ci_github_access_token
     ctx.obj["ci_job_key"] = ci_job_key
     ctx.obj["pipeline_start_timestamp"] = pipeline_start_timestamp
-    ctx.obj["dagger_logs_path"] = Path(f"{LOCAL_REPORTS_PATH_ROOT}/dagger.log") if not show_dagger_logs else None
+    ctx.obj["report_output_prefix"] = render_report_output_prefix(ctx)
+    if not show_dagger_logs:
+        dagger_log_dir = Path(f"{LOCAL_REPORTS_PATH_ROOT}/{ctx.obj['report_output_prefix']}")
+        dagger_log_dir.mkdir(parents=True, exist_ok=True)
+        dagger_log_path = Path(f"{dagger_log_dir}/dagger.log").resolve()
+        dagger_log_path.touch()
+        ctx.obj["dagger_logs_path"] = dagger_log_path
+        main_logger.info(f"Saving dagger logs to: {dagger_log_path}")
+    else:
+        ctx.obj["dagger_logs_path"] = None
 
     if pull_request_number and ci_github_access_token:
         ctx.obj["pull_request"] = github.get_pull_request(pull_request_number, ci_github_access_token)
@@ -125,6 +183,20 @@ def airbyte_ci(
         main_logger.info(f"Pull Request Number: {pull_request_number}")
         main_logger.info(f"Pipeline Start Timestamp: {pipeline_start_timestamp}")
         main_logger.info(f"Modified Files: {ctx.obj['modified_files']}")
+
+    @ctx.call_on_close
+    def upload_dagger_logs() -> None:
+        if ctx.obj["dagger_logs_path"]:
+            if ctx.obj["is_local"]:
+                main_logger.info(f"Dagger logs saved to {ctx.obj['dagger_logs_path']}")
+            if ctx.obj["is_ci"]:
+                dagger_logs_gcs_key = f"{ctx.obj['report_output_prefix']}/dagger-logs.txt"
+                upload_to_gcs(
+                    ctx.obj["dagger_logs_path"], ctx.obj["ci_report_bucket_name"], dagger_logs_gcs_key, ctx.obj["ci_gcs_credentials"]
+                )
+                gcs_uri = f"gs://{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
+                public_url = f"https://storage.googleapis.com/{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
+                main_logger.info(f"Dagger logs saved to {gcs_uri}. Public URL: {public_url}")
 
 
 airbyte_ci.add_command(connectors)
