+++ tools/ci_connector_ops/ci_connector_ops/pipelines/commands/groups/connectors.py
@@ -19,7 +19,7 @@ from ci_connector_ops.pipelines.github import update_global_commit_status_check_
 from ci_connector_ops.pipelines.pipelines.connectors import run_connectors_pipelines
 from ci_connector_ops.pipelines.publish import reorder_contexts, run_connector_publish_pipeline
 from ci_connector_ops.pipelines.tests import run_connector_test_pipeline
-from ci_connector_ops.pipelines.utils import DaggerPipelineCommand, get_modified_connectors, get_modified_metadata_files, upload_to_gcs
+from ci_connector_ops.pipelines.utils import DaggerPipelineCommand, get_modified_connectors, get_modified_metadata_files
 from ci_connector_ops.utils import ConnectorLanguage, console, get_all_released_connectors
 from rich.table import Table
 from rich.text import Text
@@ -47,57 +47,11 @@ def validate_environment(is_local: bool, use_remote_secrets: bool):
         )
 
 
-def render_report_output_prefix(ctx: click.Context) -> str:
-    """Render the report output prefix for any command in the Connector CLI.
-
-    The goal is to standardize the output of all logs and reports generated by the CLI
-    related to a specific command, and to a specific CI context.
-
-    Note: We cannot hoist this higher in the command hierarchy because only one level of
-    subcommands are available at the time the context is created.
-    """
-
-    git_branch = ctx.obj["git_branch"]
-    git_revision = ctx.obj["git_revision"]
-    pipeline_start_timestamp = ctx.obj["pipeline_start_timestamp"]
-    ci_context = ctx.obj["ci_context"]
-    ci_job_key = ctx.obj["ci_job_key"] if ctx.obj.get("ci_job_key") else ci_context
-
-    sanitized_branch = git_branch.replace("/", "_")
-
-    # get the command name for the current context, if a group then prepend the parent command name
-    invoked_subcommand = ctx.invoked_subcommand
-    parent_command_path = ctx.command_path.replace(" ", "/") if ctx.command_path else None
-    cmd = f"{parent_command_path}/{invoked_subcommand}" if parent_command_path else invoked_subcommand
-
-    path_values = [
-        cmd,
-        ci_job_key,
-        sanitized_branch,
-        pipeline_start_timestamp,
-        git_revision,
-    ]
-
-    # check all values are defined
-    if None in path_values:
-        raise ValueError(f"Missing value required to render the report output prefix: {path_values}")
-
-    # join all values with a slash, and convert all values to string
-    return "/".join(map(str, path_values))
-
-
 # COMMANDS
 
 
 @click.group(help="Commands related to connectors and connector acceptance tests.")
 @click.option("--use-remote-secrets", default=True)  # specific to connectors
-@click.option(
-    "--ci-gcs-credentials",
-    help="The service account to use during CI.",
-    type=click.STRING,
-    required=False,  # Not required for pre-release or local pipelines
-    envvar="GCP_GSM_CREDENTIALS",
-)
 @click.option(
     "--name", "names", multiple=True, help="Only test a specific connector. Use its technical name. e.g source-pokeapi.", type=str
 )
@@ -121,7 +75,6 @@ def render_report_output_prefix(ctx: click.Context) -> str:
 def connectors(
     ctx: click.Context,
     use_remote_secrets: bool,
-    ci_gcs_credentials: str,
     names: Tuple[str],
     languages: Tuple[ConnectorLanguage],
     release_stages: Tuple[str],
@@ -134,14 +87,12 @@ def connectors(
 
     ctx.ensure_object(dict)
     ctx.obj["use_remote_secrets"] = use_remote_secrets
-    ctx.obj["ci_gcs_credentials"] = ci_gcs_credentials
     ctx.obj["connector_names"] = names
     ctx.obj["connector_languages"] = languages
     ctx.obj["release_states"] = release_stages
     ctx.obj["modified"] = modified
     ctx.obj["concurrency"] = concurrency
     ctx.obj["execute_timeout"] = execute_timeout
-    ctx.obj["report_output_prefix"] = render_report_output_prefix(ctx)
 
     all_connectors = get_all_released_connectors()
 
@@ -214,7 +165,6 @@ def test(
             ci_context=ctx.obj.get("ci_context"),
             pull_request=ctx.obj.get("pull_request"),
             ci_gcs_credentials=ctx.obj["ci_gcs_credentials"],
-            should_save_report=True,
         )
         for connector, modified_files in ctx.obj["selected_connectors_and_files"].items()
     ]
@@ -232,22 +182,12 @@ def test(
         main_logger.error("An error occurred while running the test pipeline", exc_info=e)
         update_global_commit_status_check_for_tests(ctx.obj, "failure")
         return False
-    finally:
-        # We always want to save the dagger logs, even if the pipeline failed.
-        if ctx.obj["is_local"] and ctx.obj["dagger_logs_path"]:
-            main_logger.info(f"Dagger logs saved to {ctx.obj['dagger_logs_path']}")
+
+    @ctx.call_on_close
+    def send_commit_status_check() -> None:
         if ctx.obj["is_ci"]:
             global_success = all(connector_context.state is ContextState.SUCCESSFUL for connector_context in connectors_tests_contexts)
             update_global_commit_status_check_for_tests(ctx.obj, "success" if global_success else "failure")
-
-            if ctx.obj["dagger_logs_path"]:
-                dagger_logs_gcs_key = f"{ctx.obj['report_output_prefix']}/dagger-logs.txt"
-                upload_to_gcs(
-                    ctx.obj["dagger_logs_path"], ctx.obj["ci_report_bucket_name"], dagger_logs_gcs_key, ctx.obj["ci_gcs_credentials"]
-                )
-                gcs_uri = f"gs://{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
-                public_url = f"https://storage.googleapis.com/{ctx.obj['ci_report_bucket_name']}/{dagger_logs_gcs_key}"
-                main_logger.info(f"Dagger logs saved to {gcs_uri}. Public URL: {public_url}")
 
     # If we reach this point, it means that all the connectors have been tested so the pipeline did its job and can exit with success.
     return True
@@ -259,7 +199,7 @@ def build(ctx: click.Context) -> bool:
     main_logger.info(f"Will build the following connectors: {', '.join(ctx.obj['selected_connectors_names'])}.")
     connectors_contexts = [
         ConnectorContext(
-            pipeline_name="Build connector {connector.technical_name}",
+            pipeline_name=f"Build connector {connector.technical_name}",
             connector=connector,
             is_local=ctx.obj["is_local"],
             git_branch=ctx.obj["git_branch"],
