+++ tools/ci_connector_ops/ci_connector_ops/pipelines/pipelines/connectors.py
@@ -9,7 +9,6 @@ from typing import Callable, List, Optional
 
 import anyio
 import dagger
-from ci_connector_ops.pipelines import main_logger
 from ci_connector_ops.pipelines.actions import environments
 from ci_connector_ops.pipelines.bases import NoOpStep, Report, StepResult, StepStatus
 from ci_connector_ops.pipelines.contexts import ConnectorContext, ContextState
@@ -82,8 +81,6 @@ async def run_connectors_pipelines(
 
     default_connectors_semaphore = anyio.Semaphore(concurrency)
     dagger_logs_output = sys.stderr if not dagger_logs_path else open(dagger_logs_path, "w")
-    if not dagger_logs_path:
-        main_logger.info(f"Dagger logs will be saved in {dagger_logs_path}")
     async with dagger.Connection(Config(log_output=dagger_logs_output, execute_timeout=execute_timeout)) as dagger_client:
         # HACK: This is to get a long running dockerd service to be shared across all the connectors pipelines
         # Using the "normal" service binding leads to restart of dockerd during pipeline run that can cause corrupted docker state
